add first 2 functions for working with contacts-routes
